# api/swagger.yaml
...
swagger: "2.0"
info:
  description: |
    Portainer API is an HTTP API served by Portainer. It is used by the Portainer UI and everything you can do with the UI can be done using the HTTP API.
    Examples are available at https://gist.github.com/deviantony/77026d402366b4b43fa5918d41bc42f8
    You can find out more about Portainer at [http://portainer.io](http://portainer.io) and get some support on [Slack](http://portainer.io/slack/).

    # Authentication

    Most of the API endpoints require to be authenticated as well as some level of authorization to be used.
    Portainer API uses JSON Web Token to manage authentication and thus requires you to provide a token in the **Authorization** header of each request
    with the **Bearer** authentication mechanism.

    Example:
    ```
    Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwidXNlcm5hbWUiOiJhZG1pbiIsInJvbGUiOjEsImV4cCI6MTQ5OTM3NjE1NH0.NJ6vE8FY1WG6jsRQzfMqeatJ4vh2TWAeeYfDhP71YEE
    ```

    # Security

    Each API endpoint has an associated access policy, it is documented in the description of each endpoint.

    Different access policies are available:
    * Public access
    * Authenticated access
    * Restricted access
    * Administrator access

    ### Public access

    No authentication is required to access the endpoints with this access policy.

    ### Authenticated access

    Authentication is required to access the endpoints with this access policy.

    ### Restricted access

...

...

...

...

...

...

...

...

...
host: "portainer.domain"
basePath: "/api"
...
schemes:
...
paths:
...
  /stacks/{id}:
    get:
      tags:
      - "stacks"
      summary: "Inspect a stack"
      description: |
        Retrieve details about a stack.
        **Access policy**: restricted
      operationId: "StackInspect"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - name: "id"
        in: "path"
        description: "Stack identifier"
        required: true
        type: "integer"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Stack"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request"
        403:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/GenericError"
        404:
          description: "Stack not found"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Stack not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
    put:
      tags:
      - "stacks"
      summary: "Update a stack"
      description: |
        Update a stack.
        **Access policy**: restricted
      operationId: "StackUpdate"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - name: "id"
        in: "path"
        description: "Stack identifier"
        required: true
        type: "integer"
      - name: "endpointId"
        in: "query"
        description: "Stacks created before version 1.18.0 might not have an associated endpoint identifier. Use this \
        optional parameter to set the endpoint identifier used by the stack."
        type: "integer"
      - in: "body"
        name: "body"
        description: "Stack details"
        required: true
        schema:
          $ref: "#/definitions/StackUpdateRequest"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Stack"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request data format"
        403:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/GenericError"
        404:
          description: "Stack not found"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Stack not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
    delete:
      tags:
      - "stacks"
      summary: "Remove a stack"
      description: |
        Remove a stack.
        **Access policy**: restricted
      operationId: "StackDelete"
      security:
      - jwt: []
      parameters:
      - name: "id"
        in: "path"
        description: "Stack identifier"
        required: true
        type: "integer"
      - name: "external"
        in: "query"
        description: "Set to true to delete an external stack. Only external Swarm stacks are supported."
        type: "boolean"
      - name: "endpointId"
        in: "query"
        description: "Endpoint identifier used to remove an external stack (required when external is set to true)"
        type: "string"
      responses:
        204:
          description: "Success"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request"
        403:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/GenericError"
        404:
          description: "Stack not found"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Stack not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
...
  /tags:
    get:
      tags:
      - "tags"
      summary: "List tags"
      description: |
        List tags.
        **Access policy**: administrator
      operationId: "TagList"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/TagListResponse"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
    post:
      tags:
      - "tags"
      summary: "Create a new tag"
      description: |
        Create a new tag.
        **Access policy**: administrator
      operationId: "TagCreate"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - in: "body"
        name: "body"
        description: "Tag details"
        required: true
        schema:
          $ref: "#/definitions/TagCreateRequest"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Tag"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request data format"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "A tag with the specified name already exists"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
...
  /teams/{id}:
    get:
      tags:
      - "teams"
      summary: "Inspect a team"
      description: |
        Retrieve details about a team. Access is only available for administrator and leaders of that team.
        **Access policy**: restricted
      operationId: "TeamInspect"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - name: "id"
        in: "path"
        description: "Team identifier"
        required: true
        type: "integer"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Team"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request"
        403:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Access denied to resource"
        404:
          description: "Team not found"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Team not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
    put:
      tags:
      - "teams"
      summary: "Update a team"
      description: |
        Update a team.
        **Access policy**: administrator
      operationId: "TeamUpdate"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - name: "id"
        in: "path"
        description: "Team identifier"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Team details"
        required: true
        schema:
          $ref: "#/definitions/TeamUpdateRequest"
      responses:
        200:
          description: "Success"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request data format"
        404:
          description: "Team not found"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Team not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
    delete:
      tags:
      - "teams"
      summary: "Remove a team"
      description: |
        Remove a team.
        **Access policy**: administrator
      operationId: "TeamDelete"
      security:
      - jwt: []
      parameters:
      - name: "id"
        in: "path"
        description: "Team identifier"
        required: true
        type: "integer"
      responses:
        204:
          description: "Success"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request"
        404:
          description: "Team not found"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Team not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
...
  /team_memberships:
    get:
      tags:
      - "team_memberships"
      summary: "List team memberships"
      description: |
        List team memberships. Access is only available to administrators and team leaders.
        **Access policy**: restricted
      operationId: "TeamMembershipList"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/TeamMembershipListResponse"
        403:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Access denied to resource"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
    post:
      tags:
      - "team_memberships"
      summary: "Create a new team membership"
      description: |
        Create a new team memberships. Access is only available to administrators leaders of the associated team.
        **Access policy**: restricted
      operationId: "TeamMembershipCreate"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - in: "body"
        name: "body"
        description: "Team membership details"
        required: true
        schema:
          $ref: "#/definitions/TeamMembershipCreateRequest"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/TeamMembership"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request data format"
        403:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Access denied to resource"
        409:
          description: "Team membership already exists"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Team membership already exists for this user and team."
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
...
securityDefinitions:
  jwt:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  Tag:
    type: "object"
    properties:
      Id:
        type: "integer"
        example: 1
        description: "Tag identifier"
      Name:
        type: "string"
        example: "org/acme"
        description: "Tag name"
  Team:
    type: "object"
    properties:
      Id:
        type: "integer"
        example: 1
        description: "Team identifier"
      Name:
        type: "string"
        example: "developers"
        description: "Team name"
  TeamMembership:
    type: "object"
    properties:
      Id:
        type: "integer"
        example: 1
        description: "Membership identifier"
      UserID:
        type: "integer"
        example: 1
        description: "User identifier"
      TeamID:
        type: "integer"
        example: 1
        description: "Team identifier"
      Role:
        type: "integer"
        example: 1
        description: "Team role (1 for team leader and 2 for team member)"
...
  GenericError:
    type: "object"
    properties:
      err:
        type: "string"
        example: "Something bad happened"
        description: "Error message"
...
  TagListResponse:
    type: "array"
    items:
      $ref: "#/definitions/Tag"
  TagCreateRequest:
    type: "object"
    required:
    - "Name"
    properties:
      Name:
        type: "string"
        example: "org/acme"
        description: "Name"
...
  TeamUpdateRequest:
    type: "object"
    required:
    - "Name"
    properties:
      Name:
        type: "string"
        example: "developers"
        description: "Name"
...
  TeamMembershipCreateRequest:
    type: "object"
    required:
    - "UserID"
    - "TeamID"
    - "Role"
    properties:
      UserID:
        type: "integer"
        example: 1
        description: "User identifier"
      TeamID:
        type: "integer"
        example: 1
        description: "Team identifier"
      Role:
        type: "integer"
        example: 1
        description: "Role for the user inside the team (1 for leader and 2 for regular member)"
  TeamMembershipListResponse:
    type: "array"
    items:
      $ref: "#/definitions/TeamMembership"
...
  Stack_Env:
    properties:
      name:
        type: "string"
        example: "MYSQL_ROOT_PASSWORD"
      value:
        type: "string"
        example: "password"
...
  Stack:
    type: "object"
    properties:
      Id:
        type: "string"
        example: "myStack_jpofkc0i9uo9wtx1zesuk649w"
        description: "Stack identifier"
      Name:
        type: "string"
        example: "myStack"
        description: "Stack name"
      Type:
        type: "integer"
        example: "1"
        description: "Stack type. 1 for a Swarm stack, 2 for a Compose stack"
      EndpointID:
        type: "integer"
        example: "1"
        description: "Endpoint identifier. Reference the endpoint that will be used for deployment "
      EntryPoint:
        type: "string"
        example: "docker-compose.yml"
        description: "Path to the Stack file"
      SwarmID:
        type: "string"
        example: "jpofkc0i9uo9wtx1zesuk649w"
        description: "Cluster identifier of the Swarm cluster where the stack is deployed"
      ProjectPath:
        type: "string"
        example: "/data/compose/myStack_jpofkc0i9uo9wtx1zesuk649w"
        description: "Path on disk to the repository hosting the Stack file"
      Env:
        type: "array"
        description: "A list of environment variables used during stack deployment"
        items:
          $ref: "#/definitions/Stack_Env"
  StackUpdateRequest:
    type: "object"
    properties:
      StackFileContent:
        type: "string"
        example: "version: 3\n services:\n web:\n image:nginx"
        description: "New content of the Stack file."
      Env:
        type: "array"
        description: "A list of environment variables used during stack deployment"
        items:
          $ref: "#/definitions/Stack_Env"
      Prune:
        type: "boolean"
        example: false
        description: "Prune services that are no longer referenced (only available for Swarm stacks)"
...